{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = "{{ cookiecutter.project_description }}"
authors = ["Lorence Lim <jlorencelim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2.9"
django-allauth = "^0.45.0"
django-cors-headers = "^3.10.0"
django-environ = "^0.8.1"
django-extensions = "^3.1.3"
django-restql = "^0.15.1"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
dj-rest-auth = "^2.1.11"
drf-yasg = "^1.20.0"
Pillow = "^8.4.0"
psycopg2-binary = "^2.9.1"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
coverage = {version = "^6.1.1", extras = ["toml"]}
django-coverage-plugin = "^2.0.1"
django-stubs = "^1.9.0"
djangorestframework-stubs = "^1.4.0"
factory-boy = "^3.2.1"
ipdb = "^0.13.9"
mypy = "^0.910"
pre-commit = "^2.15.0"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-django = "^4.4.0"
pytest-factoryboy = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


# pytest
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "{{ cookiecutter.project_slug }}.settings"


# isort
[tool.isort]
line_length = 88
profile = "black"
known_django = "django"
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
multi_line_output = 3
skip_glob = "**/migrations/*.py"
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true


# coverage
[tool.coverage.run]
omit = [
    "*migrations*",
    "*tests*",
    "*__init__.py",
    "{{ cookiecutter.project_slug }}/*",
    "conftest.py",
    "manage.py",
]
plugins = ["django_coverage_plugin"]

[tool.coverage.report]
omit = [
    "*migrations*",
    "*tests*",
    "*__init__.py",
    "{{ cookiecutter.project_slug }}/*",
    "conftest.py",
    "manage.py",
]
